Fixed bug

+++ applications/models.py
@@ -418,7 +418,8 @@ class HackerApplication(
         self.contacted = False
         self.status_update_date = timezone.now()
         self.vote_set.all().delete()
-        self.acceptedresume.delete()
+        if hasattr(self, 'acceptedresume'):
+            self.acceptedresume.delete()
         self.save()
 
     def unset_dubious(self):
